Fix line feeds in vulnerabilty output when running in wide mode (backport)
Backport of b6cc01a558aa6add1be92e6aa2e6a88167dd1c0f:

-----

This commit is a FIX for #1069, thus when running in
wide mode it corrects an additional line feed which
happened sometimes.

As @dcooper16 pointed out it also cleans up the needless
if-statements in run_rc4(), run_lucky13() and run_beast().

It also inserts for wide mode lines a blank so the alignment
is not at the left border anymore (check for leftovers
needed).